split components files+fix missing cmd: earned-exp
