+++ TheOracle2/Interactions/SlashCommands/PlayerCharacterCommands.cs
@@ -59,6 +59,12 @@ public class EditPlayerPaths : InteractionModuleBase
     public async Task SetEarnedXp([Autocomplete(typeof(CharacterAutocomplete))] string character,
                                 [Summary(description: "The total amount of Xp this character has earned")][MinValue(0)] int earnedXp)
     {
+        if (!int.TryParse(character, out var Id)) return;
+        var pc = DbContext.PlayerCharacters.Find(Id);
+        pc.XpGained = earnedXp;
+        await DbContext.SaveChangesAsync();
+        await RespondAsync($"{pc.Name}'s XP was set to **{earnedXp}**. Their card will be updated the next time you click a button on the card.", ephemeral: true).ConfigureAwait(false);
+        return;
     }
 
     [SlashCommand("delete", "Delete a player character, removing them from the database and search results.")]
@@ -77,227 +83,5 @@ public class EditPlayerPaths : InteractionModuleBase
             .WithButton(GenericComponents.CancelButton())
             .WithButton("Delete", $"delete-player-{pc.Id}", style: ButtonStyle.Danger)
             .Build());
-    }
-}
-
-public class EditPlayerComponents : InteractionModuleBase<SocketInteractionContext<SocketMessageComponent>>
-{
-    private readonly ILogger<EditPlayerComponents> logger;
-
-    public EditPlayerComponents(EFContext efContext, ILogger<EditPlayerComponents> logger)
-    {
-        EfContext = efContext;
-        this.logger = logger;
-    }
-
-    public EFContext EfContext { get; }
-
-    [ComponentInteraction("delete-player-*")]
-    public async Task DeletePlayer(string pcId)
-    {
-        await DeferAsync();
-
-        if (!int.TryParse(pcId, out var id)) return;
-        var pc = await EfContext.PlayerCharacters.FindAsync(id);
-
-        if (pc == null)
-        {
-            await RespondAsync($"I couldn't find that character, is it maybe already deleted?", ephemeral: true);
-            return;
-        }
-
-        if (pc.DiscordGuildId != Context.Guild.Id || (pc.UserId != Context.User.Id && Context.Guild.OwnerId != Context.User.Id))
-        {
-            await RespondAsync($"You are not allowed to delete this player character.", ephemeral: true);
-            return;
-        }
-
-        var errors = new List<string>();
-        try
-        {
-            EfContext.PlayerCharacters.Remove(pc);
-            await EfContext.SaveChangesAsync();
-        }
-        catch (Exception ex)
-        {
-            logger.LogError($"Unable to delete player from database. Id: {pcId}\n{ex}");
-            await FollowupAsync($"Unable to delete player {pc.Name} from the player database. Please try again, then make an issue on github, or post on the Ironsworn discord.", ephemeral: true);
-            return;
-        }
-
-        if (pc.MessageId > 0)
-        {
-            try
-            {
-                var msg = await ((await Context.Client.GetChannelAsync(pc.ChannelId)) as IMessageChannel)?.GetMessageAsync(pc.MessageId);
-                await msg?.DeleteAsync(); //The message could've been deleted by the user.
-            }
-            catch (Exception ex)
-            {
-                logger.LogWarning($"Unable to delete player card post for player id {pcId}:\n{ex}");
-                errors.Add($"`The player was deleted from the database, but I was unable to delete player card post. If you already deleted the post you can ignore this message, otherwise you might want to delete the post yourself.`");
-            }
-        }
-        else
-        {
-            errors.Add("`Unable to delete player card post (this does not mean the character wasn't removed from command search results)`");
-        }
-
-        await Context.Interaction.Message.DeleteAsync();
-
-        string message = (errors.Count == 0) ? $"Deleted {pc.Name}" : $"Finished with error(s):\n{string.Join('\n', errors)}";
-        await FollowupAsync(message, ephemeral: true);
-    }
-}
-
-
-public class PlayerCardComponents : InteractionModuleBase<SocketInteractionContext<SocketMessageComponent>>
-{
-    public PlayerCardComponents(EFContext dbContext)
-    {
-        DbContext = dbContext;
-    }
-    public GuildPlayer GuildPlayer => GuildPlayer.AddIfMissing(Context, DbContext);
-    public EFContext DbContext { get; }
-    public override void AfterExecute(ICommandInfo command)
-    {
-        DbContext.SaveChanges();
-    }
-    [ComponentInteraction("add-momentum-*")]
-    public async Task AddMomentum(string pcId)
-    {
-        await UpdatePCValue(pcId, pc => pc.Momentum++).ConfigureAwait(false);
-    }
-
-    [ComponentInteraction("lose-momentum-*")]
-    public async Task LoseMomentum(string pcId)
-    {
-        await UpdatePCValue(pcId, pc => pc.Momentum--).ConfigureAwait(false);
-    }
-
-    [ComponentInteraction("lose-supply-*")]
-    public async Task loseSupply(string pcId)
-    {
-        await UpdatePCValue(pcId, pc => pc.Supply--).ConfigureAwait(false);
-    }
-
-    [ComponentInteraction("add-supply-*")]
-    public async Task addSupply(string pcId)
-    {
-        await UpdatePCValue(pcId, pc => pc.Supply++).ConfigureAwait(false);
-    }
-
-    [ComponentInteraction("lose-health-*")]
-    public async Task loseHealth(string pcId)
-    {
-        await UpdatePCValue(pcId, pc => pc.Health--).ConfigureAwait(false);
-    }
-
-    [ComponentInteraction("add-health-*")]
-    public async Task addHealth(string pcId)
-    {
-        await UpdatePCValue(pcId, pc => pc.Health++).ConfigureAwait(false);
-    }
-
-    [ComponentInteraction("lose-spirit-*")]
-    public async Task loseSpirit(string pcId)
-    {
-        await UpdatePCValue(pcId, pc => pc.Spirit--).ConfigureAwait(false);
-    }
-
-    [ComponentInteraction("add-spirit-*")]
-    public async Task addSpirit(string pcId)
-    {
-        await UpdatePCValue(pcId, pc => pc.Spirit++).ConfigureAwait(false);
-    }
-
-    [ComponentInteraction("add-xp-*")]
-    public async Task addXp(string pcId)
-    {
-        await UpdatePCValue(pcId, pc => pc.XpGained++).ConfigureAwait(false);
-    }
-
-    [ComponentInteraction("lose-xp-*")]
-    public async Task loseXp(string pcId)
-    {
-        await UpdatePCValue(pcId, pc => pc.XpGained--).ConfigureAwait(false);
-    }
-
-    [ComponentInteraction("burn-momentum-*")]
-    public async Task burn(string pcId)
-    {
-        await UpdatePCValue(pcId, pc => pc.BurnMomentum()).ConfigureAwait(false);
-    }
-
-    [ComponentInteraction("player-more-*")]
-    public async Task toggleMore(string pcId)
-    {
-        try
-        {
-            await Context.Interaction.UpdateAsync(msg =>
-            {
-                ComponentBuilder components = ComponentBuilder.FromMessage(Context.Interaction.Message);
-
-                components.TryAdd(ButtonBuilder.CreateSuccessButton("+Xp", $"add-xp-{pcId}").Build(), 2);
-                components.TryAdd(ButtonBuilder.CreateSecondaryButton("-Xp", $"lose-xp-{pcId}").Build(), 2);
-
-                components.ReplaceComponentById($"player-more-{pcId}", ButtonBuilder.CreatePrimaryButton("less", $"player-less-{pcId}").Build());
-
-                msg.Components = components.Build();
-            }).ConfigureAwait(false);
-        }
-        catch (HttpException hex)
-        {
-            string json = JsonConvert.SerializeObject(hex.Errors, Formatting.Indented);
-
-            Console.WriteLine(json);
-            throw;
-        }
-    }
-
-    [ComponentInteraction("player-less-*")]
-    public async Task toggleLess(string pcId)
-    {
-        try
-        {
-            await Context.Interaction.UpdateAsync(msg =>
-            {
-                ComponentBuilder components = ComponentBuilder.FromMessage(Context.Interaction.Message)
-                .RemoveComponentById($"add-xp-{pcId}")
-                .RemoveComponentById($"lose-xp-{pcId}");
-
-                components.ReplaceComponentById($"player-less-{pcId}", ButtonBuilder.CreatePrimaryButton("...", $"player-more-{pcId}").Build());
-
-                msg.Components = components.Build();
-            }).ConfigureAwait(false);
-        }
-        catch (HttpException hex)
-        {
-            string json = JsonConvert.SerializeObject(hex.Errors, Formatting.Indented);
-            Console.WriteLine(json);
-            throw;
-        }
-    }
-
-    private async Task UpdatePCValue(string pcId, Action<PlayerCharacter> change)
-    {
-        if (!int.TryParse(pcId, out var Id)) return;
-
-        var pc = await DbContext.PlayerCharacters.FindAsync(Id);
-        if (pc.MessageId != Context.Interaction.Message.Id)
-        {
-            pc.MessageId = Context.Interaction.Message.Id;
-            pc.ChannelId = Context.Interaction.Channel.Id;
-        }
-        change(pc);
-        GuildPlayer.LastUsedPcId = Id;
-        // await DbContext.SaveChangesAsync();
-        // TODO: commenting out the above to see what breaks.
-
-        var entity = new PlayerCharacterEntity(pc);
-        await Context.Interaction.UpdateAsync(msg =>
-        {
-            msg.Embeds = entity.GetEmbeds();
-        }).ConfigureAwait(false);
     }
 }
